+++ currency_exchange/templates/base.html
@@ -57,7 +57,10 @@
                 <a class="nav-link" href="{% url 'account_logout' %}">Wallet</a>
               </li>
               <li>
-                <a class="nav-link" href="{% url 'users:profile' request.user %}">Profile</a>
+                <a class="nav-link" href="{% url 'users:profile' request.user.username %}">Profile</a>
+              </li>
+              <li>
+                <a class="nav-link" href="{% url 'users:detail' request.user.username %}">Detail</a>
               </li>
               <li>
                 {# URL provided by django-allauth/account/urls.py #}
