+++ currency_exchange/templates/users/user_form.html
@@ -5,7 +5,7 @@
 
 {% block content %}
   <h1>Hi {{ user.username }}</h1>
-  <form class="form-horizontal" method="post" action="{% url 'users:profile'  %}">
+  <form class="form-horizontal" method="post" action="{% url 'users:update'  %}">
     {% csrf_token %}
     {{ form|crispy }}
     <div class="control-group">
